Extracted SubNavigation code to its own component
